merges categorycontainer and categorylist
